move pdf related methods into separated module
